requirements.txt
django
djangorestframework
pyodbc
django-mssql-backend
python-dotenv
django-cors-headers
djangorestframework-simplejwt
spacy
mysqlclient